:zap: feat(Methods):
Go does not have classes. However, you can define methods on types. A
method is a function with a special receiver argument. The receiver
appears in its own argument list between the func keyword and the method
name. In this example, the Abs method has a receiver of type Vertex name
v.